Refactor Discord bot logic and enhance response handling
- Introduced new regex patterns for detecting small talk, corny jokes, and meta leaks.
- Added functions to identify and handle specific user interactions, including GIF requests and inside jokes.
- Updated the prompt generation to improve response behavior and tone control.
- Modified the style profile to use integers for average character counts.
- Created a new module for handling Tenor GIF links, including loading from CSV and checking for GIF requests.
- Adjusted the preprocessing of CSV data to disable URL replacement.
- Enhanced the bot's response logic to avoid echoing user messages and using stock phrases.

+++ musinksbot/style_profile.py
@@ -29,7 +29,7 @@ DEFAULT_MARKERS = [
 @dataclass(frozen=True)
 class StyleProfile:
     messages: int
-    avg_chars: float
+    avg_chars: int
     p50_chars: float
     p90_chars: float
     markers: list[str]
